+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The `pycldf` package adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+Fixed bug whereby `dict` returned by `orm.Language.as_geojson_feature` could not be serialized
+by `json.dumps`.
+
+
 ## [1.37.1] - 2024-03-18
 
 Fixed bug whereby component names where the CSV filenames contain underscores were not translated