+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 The `pycldf` package adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [1.38.0] - 2024-04-26
 
 - Fixed bug whereby `dict` returned by `orm.Language.as_geojson_feature` could not be serialized
   by `json.dumps`.